requirements.txt
pytest~=8.3.4
pytest-asyncio~=0.25.2
fastapi~=0.115.8
httpx~=0.27.2
uvicorn~=0.32.1
starlette~=0.45.3
python-jose~=3.3.0
pydantic~=2.10.4
asyncpg~=0.29.0
pytz~=2024.1
passlib~=1.7.4
pillow~=11.0.0
pypdf2~=3.0.1
boto3~=1.34.154
pytesseract~=0.3.13
torch~=2.6.0
pdf2image~=1.17.0
transformers~=4.45.2
ollama~=0.4.4
pandas~=2.2.3
camelot-py~=1.0.0
numpy~=1.26.4
requests~=2.32.3
paramiko~=3.5.0
scp~=0.14.1
rapidfuzz~=3.5.2